Separate Achievement from AchievementPanelController
