+++ Assets/Scripts/Achievment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3f78083368044e8bb1ad57e7297e87a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
